Fixes for assembly revision & android versionCode.
Assembly revision numbers have a max. value in Visual Studio which was being exceeded by XUVI towards the end of the day (it being calculated by seconds passed since midnight). Now the number is divided by two to ensure it always stays low enough.

It was assumed Android's versionCode attribute could accept decimal numbers. It appears this is not the case and that it has to be an Integer. So the revision is now excluded from that (and just the build number is used).